backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.4.1
requests==2.31.0
textblob==0.17.1
beautifulsoup4==4.12.2
tweepy==4.14.0
google-api-python-client==2.95.0
